INOVA-1-resolving conflicts and codding
