+++ lib/src/utils/colors.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+const colors = [
+  {
+    "colorName": "naviGrey",
+    "colorValue": Color(0xFFF5F5F5),
+  },
+  {
+    "colorName": "fontGrey",
+    "colorValue": Color(0xFFBFBFBF),
+  },
+  {
+    "colorName": "black",
+    "colorValue": Color(0xFF000000),
+  },
+  {
+    "colorName": "white",
+    "colorValue": Color(0xFFFFFFFF),
+  },
+  {
+    "colorName": "red",
+    "colorValue": Colors.red,
+  },
+  {
+    "colorName": "backgroundPurple",
+    "colorValue": Color(0xFF4065FC),
+  },
+  {
+    "colorName": "iconPurple",
+    "colorValue": Color(0xFF3B64FA),
+  },
+  {
+    "colorName": "lightPurplegrey",
+    "colorValue": Color(0xFFE8ECFF),
+  }
+];
